refactor: blocksci doctor
